+++ lib/invoiced.rb
@@ -35,6 +35,8 @@ require 'invoiced/note'
 require 'invoiced/payment_plan'
 require 'invoiced/payment_source_object'
 require 'invoiced/payment_source'
+require 'invoiced/bank_account'
+require 'invoiced/card'
 require 'invoiced/plan'
 require 'invoiced/subscription'
 require 'invoiced/task'
