+++ lib/invoiced/bank_account.rb
@@ -1,0 +1,7 @@
+module Invoiced
+  class BankAccount < PaymentSourceObject
+    include Invoiced::Operations::Delete
+
+    OBJECT_NAME = 'bank_account'
+  end
+end
