+++ lib/invoiced/card.rb
@@ -1,0 +1,7 @@
+module Invoiced
+  class Card < PaymentSourceObject
+    include Invoiced::Operations::Delete
+
+    OBJECT_NAME = 'card'
+  end
+end
