+++ lib/invoiced/payment_source.rb
@@ -5,16 +5,4 @@ module Invoiced
 
     OBJECT_NAME = 'payment_source'
   end
-
-  class BankAccount < PaymentSourceObject
-    include Invoiced::Operations::Delete
-
-    OBJECT_NAME = 'bank_account'
-  end
-
-  class Card < PaymentSourceObject
-    include Invoiced::Operations::Delete
-
-    OBJECT_NAME = 'card'
-  end
 end
